arch: Restructure packages
- It removes `shared` package - no sense in GRE2G context.
- Renames `config.py` to `hydra_config.py`.
- Fixes `GetHydraConfig` applied on a class
when the method is called from a class object.
